DeliveryService criado e implementado

+++ src/main/java/com/algaworks/algalog/domain/service/ClientService.java
@@ -3,6 +3,7 @@ package com.algaworks.algalog.domain.service;
 import com.algaworks.algalog.domain.dto.ClientDTO;
 import com.algaworks.algalog.domain.dto.UpdateClientDTO;
 import com.algaworks.algalog.domain.entity.Client;
+import com.algaworks.algalog.domain.exception.EmailUsedExecption;
 import com.algaworks.algalog.domain.reposiotry.ClientRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
@@ -16,7 +17,9 @@ public class ClientService {
 
     @Transactional
     public Client save(ClientDTO client) {
-        return clientRepository.save(new Client(null, client.name(), client.email(), client.fone()));
+        if(!clientRepository.existsByEmail(client.email())){
+            return clientRepository.save(new Client(null, client.name(), client.email(), client.fone()));
+        }else throw new EmailUsedExecption("This email is already in use.");
     }
 
     @Transactional
@@ -31,7 +34,9 @@ public class ClientService {
 
     @Transactional
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (clientRepository.existsById(id)) {
+            clientRepository.deleteById(id);
+        } else throw new EntityNotFoundException();
     }
 
 }
